etl/postgres_to_es/sqlite_to_postgres/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
attrs==23.1.0
click==8.1.3
colorama==0.4.6
Django==3.2.4
django-debug-toolbar==3.4.0
django-extensions==3.2.1
exceptiongroup==1.1.1
Faker==18.9.0
flake8==4.0.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.6.1
mypy-extensions==1.0.0
oitnb==0.2.2
packaging==23.1
pathspec==0.11.1
pluggy==1.0.0
psycopg2-binary==2.9
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.6.2
